+++ src/main/java/com/froggylord/constellation/config/LyraConfig.java
@@ -170,6 +170,10 @@ public class LyraConfig extends BaseConfigGroup {
     public int profileViewerInventoryRows = 5;
     public int profileViewerInventorySlotColor = 0xD020202C;
     public int profileViewerInventoryHoverColor = 0xE03C3C55;
+    public boolean profileWealth = true;
+    public int profileWealthItemsPerCategory = 8;
+    public int profileWealthMaxPriceRequests = 30;
+    public int profileWealthRequestIntervalMs = 1000;
     public boolean auctionHelper = true;
     public boolean auctionOutbidAlert = true;
     public boolean auctionOutbidTitle = true;